us-stock-assistant/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.12.5
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.7
python-multipart>=0.0.6
redis>=5.0.1
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-core>=0.2.2
langgraph>=0.0.69
pandas>=2.1.4
openpyxl>=3.1.2
slowapi>=0.1.9
apscheduler>=3.10.4
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.27.0
hypothesis>=6.92.1
fakeredis>=2.21.0
mcp>=1.0.0
prometheus-client>=0.19.0
transformers==4.40.0
torch>=2.2.0,<2.4.0
